requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/base.in
#
babel==2.9.1
    # via -r requirements/base.in
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.6
    # via requests
click==8.0.1
    # via flask
feedgen==0.9.0
    # via -r requirements/base.in
flask==2.0.3
    # via
    #   -r requirements/base.in
    #   flask-assets
    #   flask-caching
    #   flask-flatpages
    #   flask-sqlalchemy
flask-assets==2.0
    # via -r requirements/base.in
flask-caching==1.10.1
    # via -r requirements/base.in
flask-flatpages==0.7.3
    # via -r requirements/base.in
flask-sqlalchemy==2.5.1
    # via -r requirements/base.in
greenlet==1.1.1
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements/base.in
idna==3.2
    # via requests
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via
    #   flask
    #   flask-flatpages
lxml==4.6.5
    # via feedgen
markdown==3.3.4
    # via flask-flatpages
markupsafe==2.0.1
    # via jinja2
pyscss==1.3.7
    # via -r requirements/base.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.in
    #   feedgen
python-dotenv==0.19.0
    # via -r requirements/base.in
pytz==2021.1
    # via
    #   -r requirements/base.in
    #   babel
pyyaml==5.4.1
    # via flask-flatpages
requests==2.26.0
    # via -r requirements/base.in
six==1.16.0
    # via
    #   flask-flatpages
    #   pyscss
    #   python-dateutil
sqlalchemy==1.4.23
    # via flask-sqlalchemy
urllib3==1.26.6
    # via requests
webassets==2.0
    # via flask-assets
werkzeug==2.0.1
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
